LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/container-queries/anchored-fallback-basic-expected.txt


FAIL @container anchored() without applied fallback assert_equals: expected "no" but got ""
FAIL @container anchored() with fallback applied assert_equals: expected "yes" but got ""
